Refactor to split files to make server code cleaner
