+++ README.md
@@ -10,7 +10,8 @@ Attempts to implement BFF (backend for frontend) pattern for login using oidc. I
 ### Prerequisites
 1. Set up google oidc client at https://console.cloud.google.com/apis/credentials
 2. Get GOOGLE_CLIENT_ID and GOOGLE_CLIENT_SECRET and set them as environment variables before you run the server binary
-3. Make sure to set authorised redirect uri to http://localhost:8080/api/token_exchange
+3. Set an environment variable SERVER_SECRET_KEY to a random string with at least 64 characters for cryptographically signing your tokens. Keep this a secret.
+4. Make sure to set authorised redirect uri to http://localhost:8080/api/token_exchange in your google oauth app
 
 ### Instructions
 1. Build frontend distribution using Trunk
