+++ server/Cargo.toml
@@ -12,6 +12,7 @@ env_logger = "0.9.0"
 log = "0.4"
 actix-files = "0.6.2"
 actix-session = {version = "0.7.2", features = ["cookie-session"]}
+tokio = {version = "1.20.0", features = ["full"]}
 openidconnect = "2.3.2"
 serde = { version = "1.0", features = ["derive"] }
